requirements.txt
dask>=2021.8.0
matplotlib>=3.5.0
netCDF4>=1.5.7
numpy>=1.20.0
pandas>=1.3.0
pyyaml>=6.0
scipy>=1.7.0
xarray>=0.19.0
